add assign utility
