+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './coercions'
 export * from './refinements'
 export * from './struct'
 export * from './types'
+export * from './utilities'
