+++ src/types.ts
@@ -1,5 +1,6 @@
-import { Struct, StructType, coerce, StructContext } from './struct'
-import { TupleSchema, ObjectSchema, InferObjectStruct } from './utils'
+import { Struct, StructType, coerce } from './struct'
+import { TupleSchema, ObjectSchema, InferObjectStruct } from './xtras'
+import { struct } from './utilities'
 
 /**
  * Validate any value.
@@ -59,18 +60,6 @@ export function boolean(): Struct<boolean> {
 export function date(): Struct<Date> {
   return struct('Date', (value) => {
     return value instanceof Date && !isNaN(value.getTime())
-  })
-}
-
-/**
- * Validate that a value dynamically, determing which struct to use at runtime.
- */
-
-export function dynamic<T>(
-  fn: (value: unknown, ctx: StructContext) => Struct<T>
-): Struct<T> {
-  return struct('Dynamic<...>', (value, ctx) => {
-    return ctx.check(value, fn(value, ctx))
   })
 }
 
@@ -178,24 +167,6 @@ export function intersection(Structs: Struct<any>[]): any {
 }
 
 /**
- * Validate a value lazily, by constructing the struct right before the first
- * validation. This is useful for cases where you want to have self-referential
- * structs for nested data structures.
- */
-
-export function lazy<T>(fn: () => Struct<T>): Struct<T> {
-  let S: Struct<T> | undefined
-
-  return struct('Lazy<...>', (value, ctx) => {
-    if (!S) {
-      S = fn()
-    }
-
-    return ctx.check(value, S)
-  })
-}
-
-/**
  * Validate that a value is a specific constant.
  */
 
@@ -298,24 +269,6 @@ export function object<S extends ObjectSchema>(schema?: S): any {
 }
 
 /**
- * Validates that a value is an object without a subset of properties.
- */
-
-export function omit<S extends ObjectSchema, K extends keyof S>(
-  struct: InferObjectStruct<S>,
-  keys: K[]
-): InferObjectStruct<Omit<S, K>> {
-  const { schema } = struct
-  const subschema: any = { ...schema }
-
-  for (const key of keys) {
-    delete subschema[key]
-  }
-
-  return object(subschema as Omit<S, K>)
-}
-
-/**
  * Augment a struct to make it optionally accept `undefined` values.
  */
 
@@ -327,41 +280,6 @@ export function optional<T, S>(struct: Struct<T, S>): Struct<T | undefined, S> {
       return value === undefined || ctx.check(value, struct)
     },
   })
-}
-
-/**
- * Validate that a partial object with specific entry values.
- */
-
-export function partial<S extends ObjectSchema>(
-  struct: InferObjectStruct<S> | S
-): InferObjectStruct<{ [K in keyof S]: Struct<StructType<S[K]> | undefined> }> {
-  const schema: any =
-    struct instanceof Struct ? { ...struct.schema } : { ...struct }
-
-  for (const key in schema) {
-    schema[key] = optional(schema[key])
-  }
-
-  return object(schema) as any
-}
-
-/**
- * Validates that a value is an object with a subset of properties.
- */
-
-export function pick<S extends ObjectSchema, K extends keyof S>(
-  struct: InferObjectStruct<S>,
-  keys: K[]
-): InferObjectStruct<Pick<S, K>> {
-  const { schema } = struct
-  const subschema: any = {}
-
-  for (const key of keys) {
-    subschema[key] = schema[key]
-  }
-
-  return object(subschema as Pick<S, K>)
 }
 
 /**
@@ -416,17 +334,6 @@ export function string(): Struct<string> {
   return struct('string', (value) => {
     return typeof value === 'string'
   })
-}
-
-/**
- * Define a `Struct` instance with a type and validation function.
- */
-
-export function struct<T>(
-  name: string,
-  validator: Struct<T>['validator']
-): Struct<T, null> {
-  return new Struct({ type: name, validator, schema: null })
 }
 
 /**
